+++ app/api/ada/transactions/transfer/legacyYoroi.test.js
@@ -130,6 +130,7 @@ describe('Haskell Shelley era tx format tests', () => {
     expect(transferInfo.receiver).toBe(outAddress);
 
     // check tx itself
+    if (!transferInfo.encodedTx) throw new Error(`Tx not signed`);
     const signedTx = RustModule.WalletV4.Transaction.from_bytes(transferInfo.encodedTx);
     const body = signedTx.body();
     expect(body.inputs().len()).toBe(1);
