--- app/components/wallet/staking/WithdrawalTxDialog.scss
@@ -1,149 +1,0 @@
-@import '../../themes/mixins/error-message';
-@import '../../themes/mixins/loading-spinner';
-
-.body {
-  color: var(--theme-bordered-box-text-color);
-  font-family: var(--font-regular);
-  font-size: 14px;
-  line-height: 19px;
-  
-  .currencySymbol {
-    font-family: var(--font-light);
-    text-transform: uppercase;
-  }
-
-  .addressLabelWrapper,
-  .amountFeesWrapper {
-    margin-bottom: 20px;
-  }
-
-  .addressLabel {
-    font-size: 12px;
-    font-weight: 600;
-    letter-spacing: 0.5px;
-    line-height: 1.38;
-  }
-
-  .addressSubLabel {
-    margin-top: 6px;
-  }
-
-  .address {
-    line-height: 1.38;
-  }
-
-  .amountFeesWrapper {
-    display: flex;
-    flex-direction: row;
-
-    .amountWrapper,
-    .feesWrapper {
-      width: 50%;
-    }
-
-    .amountWrapper {
-      margin-right: 10px;
-    }
-
-    .feesWrapper {
-      margin-left: 10px;
-    }
-
-    .amountLabel,
-    .feesLabel {
-      font-size: 12px;
-      font-weight: 600;
-      letter-spacing: 0.5px;
-      line-height: 1.38;
-      text-transform: uppercase;
-    }
-
-    .amount,
-    .fees {
-      color: var(--theme-send-confirmation-dialog-send-values-color);
-      line-height: 1.38;
-      margin-top: 6px;
-      word-break: break-word;
-    }
-
-    .fees {
-      opacity: 0.5;
-    }
-
-    .amountSmall {
-      @extend .amount;
-      font-size: 12px;
-    }
-
-    .feesSmall {
-      @extend .fees;
-      font-size: 12px;
-    }
-  }
-
-  .totalAmountLabel {
-    font-size: 12px;
-    font-weight: 600;
-    letter-spacing: 0.5px;
-    line-height: 1.38;
-    text-transform: uppercase;
-  }
-
-  .totalAmount {
-    color: var(--theme-send-confirmation-dialog-send-values-color);
-    line-height: 1.38;
-    margin-top: 6px;
-    word-break: break-word;
-  }
-
-  .totalAmountSmall {
-    @extend .totalAmount;
-    font-size: 12px;
-  }
-
-  .form {
-    margin-top: 20px;
-  }
-
-  .errorWrapper {
-    margin: 1% 0%;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-  }
-
-  .error {
-    @include error-message;
-    font-size: 16px;
-    line-height: 22px;
-    margin-bottom: 12px;
-    word-break: break-word;
-    text-align: center;
-  }
-}
-
-:global(.YoroiClassic) .body {
-  .addressLabel {
-    text-transform: uppercase;
-  }
-}
-
-:global(.YoroiModern) .body {
-  .addressLabel {
-    font-size: 16px;
-    text-transform: capitalize;
-  }
-
-  .amountFeesWrapper {
-    .amountLabel,
-    .feesLabel {
-      font-size: 16px;
-      text-transform: capitalize;
-    }
-  }
-
-  .totalAmountLabel {
-    font-size: 16px;
-    text-transform: capitalize;
-  }
-}
