+++ app/containers/settings/categories/GeneralSettingsPage.stories.js
@@ -8,7 +8,8 @@ import { action } from '@storybook/addon-actions';
 import { LANGUAGES } from '../../../i18n/translations';
 import GeneralSettingsPage from './GeneralSettingsPage';
 import { withScreenshot } from 'storycap';
-import { globalKnobs, walletLookup } from '../../../../stories/helpers/StoryWrapper';
+import { globalKnobs, } from '../../../../stories/helpers/StoryWrapper';
+import { walletLookup } from '../../../../stories/helpers/WalletCache';
 import { getVarsForTheme } from '../../../stores/toplevel/ProfileStore';
 import { wrapSettings } from '../../../Routes';
 import { mockSettingsProps } from '../Settings.mock';
