+++ app/containers/transfer/UnmangleTxDialogContainer.js
@@ -2,34 +2,21 @@
 import React, { Component } from 'react';
 import type { Node } from 'react';
 import { observer } from 'mobx-react';
-import { action, computed, observable } from 'mobx';
-import { intlShape, } from 'react-intl';
-import ReactToolboxMobxForm from '../../utils/ReactToolboxMobxForm';
+import { computed, } from 'mobx';
 import type { InjectedOrGenerated } from '../../types/injectedPropsType';
-import LegacyTransferLayout from '../../components/transfer/LegacyTransferLayout';
-import TransferSummaryPage from '../../components/transfer/TransferSummaryPage';
-import YoroiTransferErrorPage from './YoroiTransferErrorPage';
-import VerticallyCenteredLayout from '../../components/layout/VerticallyCenteredLayout';
-import Dialog from '../../components/widgets/Dialog';
-import LoadingSpinner from '../../components/widgets/LoadingSpinner';
-import { formattedWalletAmount } from '../../utils/formatters';
 import {
   asHasUtxoChains,
 } from '../../api/ada/lib/storage/models/PublicDeriver/traits';
-import SpendingPasswordInput from '../../components/widgets/forms/SpendingPasswordInput';
-import { addressToDisplayString, getAddressPayload } from '../../api/ada/lib/storage/bridge/utils';
-import globalMessages from '../../i18n/global-messages';
+import { getAddressPayload } from '../../api/ada/lib/storage/bridge/utils';
 import type { ConfigType } from '../../../config/config-types';
-import type { $npm$ReactIntl$IntlFormat } from 'react-intl';
 import { PublicDeriver } from '../../api/ada/lib/storage/models/PublicDeriver/index';
-import { SelectedExplorer } from '../../domain/SelectedExplorer';
-import type { UnitOfAccountSettingType } from '../../types/unitOfAccountType';
 import LocalizableError from '../../i18n/LocalizableError';
 import type { SetupSelfTxRequest } from '../../stores/toplevel/TransactionBuilderStore';
-import { ApiOptions, getApiForNetwork, getApiMeta } from '../../api/common/utils';
 import { GROUP_MANGLED } from '../../stores/stateless/addressStores';
 import type { IAddressTypeStore, IAddressTypeUiSubset } from '../../stores/stateless/addressStores';
 import type { ISignRequest } from '../../api/common/lib/transactions/ISignRequest';
+import TransferSendPage from './TransferSendPage';
+import type { GeneratedData as TransferSendData } from './TransferSendPage';
 
 declare var CONFIG: ConfigType;
 
@@ -42,17 +29,6 @@ type Props = {|
 
 @observer
 export default class UnmangleTxDialogContainer extends Component<Props> {
-
-  static contextTypes: {|intl: $npm$ReactIntl$IntlFormat|} = {
-    intl: intlShape.isRequired,
-  };
-
-  @observable spendingPasswordForm: void | ReactToolboxMobxForm;
-
-  @action
-  setSpendingPasswordForm(form: ReactToolboxMobxForm) {
-    this.spendingPasswordForm = form;
-  }
 
   componentDidMount() {
     const selected = this.generated.stores.wallets.selected;
@@ -87,164 +63,38 @@ export default class UnmangleTxDialogContainer extends Component<Props> {
     });
   }
 
-  componentWillUnmount() {
-    const builderActions = this.generated.actions.txBuilderActions;
-    builderActions.reset.trigger();
-    this.generated.stores.wallets.sendMoneyRequest.reset();
-  }
-
-  submit: void => Promise<void> = async () => {
-    const selected = this.generated.stores.wallets.selected;
-    if (selected == null) {
-      throw new Error(`${nameof(UnmangleTxDialogContainer)} no wallet selected`);
-    }
-    if (this.spendingPasswordForm == null) {
-      throw new Error(`${nameof(UnmangleTxDialogContainer)} form not set`);
-    }
-    this.spendingPasswordForm.submit({
-      onSuccess: async (form) => {
-        const { walletPassword } = form.values();
-
-        const txBuilderStore = this.generated.stores.transactionBuilderStore;
-        if (txBuilderStore.tentativeTx == null) return;
-        await this.generated.actions.wallets.sendMoney.trigger({
-          signRequest: txBuilderStore.tentativeTx,
-          password: walletPassword,
-          publicDeriver: selected,
-        });
-      },
-      onError: () => {}
-    });
-  };
-
   render(): Node {
     const txBuilder = this.generated.stores.transactionBuilderStore;
-
-    if (txBuilder.setupSelfTx.error != null) {
-      return (
-        <YoroiTransferErrorPage
-          error={txBuilder.setupSelfTx.error}
-          onCancel={this.props.onClose}
-          classicTheme={this.generated.stores.profile.isClassicTheme}
-        />
-      );
-    }
-
-    if (txBuilder.tentativeTx == null) {
-      return this.getSpinner();
-    }
-    const tentativeTx = txBuilder.tentativeTx;
-    return this.getContent(tentativeTx);
-  }
-
-  getSpinner: void => Node = () => {
-    const { intl } = this.context;
     return (
-      <Dialog
-        title={intl.formatMessage(globalMessages.processingLabel)}
-        closeOnOverlayClick={false}
-      >
-        <LegacyTransferLayout>
-          <VerticallyCenteredLayout>
-            <LoadingSpinner />
-          </VerticallyCenteredLayout>
-        </LegacyTransferLayout>
-      </Dialog>
-    );
-  }
-
-  getContent:  ISignRequest<any> => Node = (
-    tentativeTx
-  ) => {
-    const selected = this.generated.stores.wallets.selected;
-    if (selected == null) {
-      throw new Error(`${nameof(UnmangleTxDialogContainer)} no wallet selected`);
-    }
-    const api = getApiForNetwork(selected.getParent().getNetworkInfo());
-    if (api !== ApiOptions.ada) {
-      throw new Error(`${nameof(UnmangleTxDialogContainer)} not ADA API type`);
-    }
-    const apiMeta = getApiMeta(api);
-    if (apiMeta == null) throw new Error(`${nameof(UnmangleTxDialogContainer)} no API selected`);
-
-    const coinPrice: ?number = this.generated.stores.profile.unitOfAccount.enabled
-      ? (
-        this.generated.stores.coinPriceStore.getCurrentPrice(
-          apiMeta.meta.primaryTicker,
-          this.generated.stores.profile.unitOfAccount.currency
-        )
-      )
-      : null;
-
-    const transferTx = {
-      recoveredBalance: tentativeTx.totalInput(true),
-      fee: tentativeTx.fee(true),
-      senders: tentativeTx
-        .uniqueSenderAddresses(),
-      receiver: tentativeTx
-        .receivers(false)[0],
-    };
-
-    const spendingPasswordForm = (<SpendingPasswordInput
-      setForm={(form) => this.setSpendingPasswordForm(form)}
-      classicTheme={this.generated.stores.profile.isClassicTheme}
-      isSubmitting={this.generated.stores.wallets.sendMoneyRequest.isExecuting}
-    />);
-
-    const { intl } = this.context;
-
-    return (
-      <TransferSummaryPage
-        form={spendingPasswordForm}
-        formattedWalletAmount={amount => formattedWalletAmount(
-          amount,
-          apiMeta.meta.decimalPlaces.toNumber(),
-        )}
-        selectedExplorer={this.generated.stores.explorers.selectedExplorer
-          .get(selected.getParent().getNetworkInfo().NetworkId) ?? (() => { throw new Error('No explorer for wallet network'); })()
-        }
-        transferTx={transferTx}
-        onSubmit={this.submit}
-        isSubmitting={this.generated.stores.wallets.sendMoneyRequest.isExecuting}
-        onCancel={this.props.onClose}
-        error={this.generated.stores.wallets.sendMoneyRequest.error}
-        dialogTitle={intl.formatMessage(globalMessages.walletSendConfirmationDialogTitle)}
-        coinPrice={coinPrice}
-        unitOfAccountSetting={this.generated.stores.profile.unitOfAccount}
-        addressToDisplayString={
-          addr => addressToDisplayString(addr, selected.getParent().getNetworkInfo())
-        }
+      <TransferSendPage
+        {...this.generated.TransferSendProps}
+        onClose={this.props.onClose}
+        transactionRequest={{
+          error: txBuilder.setupSelfTx.error,
+          result: txBuilder.tentativeTx,
+          reset: this.generated.actions.txBuilderActions.reset.trigger,
+        }}
+        toTransferTx={tentativeTx => ({
+          recoveredBalance: tentativeTx.totalInput(true),
+          fee: tentativeTx.fee(true),
+          senders: tentativeTx
+            .uniqueSenderAddresses(),
+          receiver: tentativeTx
+            .receivers(false)[0],
+        })}
       />
     );
   }
 
   @computed get generated(): {|
+    TransferSendProps: InjectedOrGenerated<TransferSendData>,
     actions: {|
       txBuilderActions: {|
         reset: {| trigger: (params: void) => void |},
         initialize: {| trigger: (params: SetupSelfTxRequest) => Promise<void> |},
       |},
-      wallets: {|
-        sendMoney: {|
-          trigger: (params: {|
-            password: string,
-            publicDeriver: PublicDeriver<>,
-            signRequest: ISignRequest<any>,
-          |}) => Promise<void>
-        |}
-      |},
     |},
     stores: {|
-      coinPriceStore: {|
-        getCurrentPrice: (from: string, to: string) => ?number
-      |},
-      explorers: {|
-        selectedExplorer: Map<number, SelectedExplorer>,
-      |},
-      profile: {|
-        isClassicTheme: boolean,
-        unitOfAccount: UnitOfAccountSettingType
-      |},
       addresses: {|
         addressSubgroupMap: $ReadOnlyMap<Class<IAddressTypeStore>, IAddressTypeUiSubset>,
       |},
@@ -256,11 +106,6 @@ export default class UnmangleTxDialogContainer extends Component<Props> {
       |},
       wallets: {|
         selected: null | PublicDeriver<>,
-        sendMoneyRequest: {|
-          error: ?LocalizableError,
-          isExecuting: boolean,
-          reset: () => void
-        |},
       |}
     |}
     |} {
@@ -272,24 +117,10 @@ export default class UnmangleTxDialogContainer extends Component<Props> {
     }
     const { stores, actions } = this.props;
     return Object.freeze({
+      TransferSendProps: ({ actions, stores, }: InjectedOrGenerated<TransferSendData>),
       stores: {
-        explorers: {
-          selectedExplorer: stores.explorers.selectedExplorer,
-        },
-        profile: {
-          isClassicTheme: stores.profile.isClassicTheme,
-          unitOfAccount: stores.profile.unitOfAccount,
-        },
         wallets: {
           selected: stores.wallets.selected,
-          sendMoneyRequest: {
-            reset: stores.wallets.sendMoneyRequest.reset,
-            error: stores.wallets.sendMoneyRequest.error,
-            isExecuting: stores.wallets.sendMoneyRequest.isExecuting,
-          },
-        },
-        coinPriceStore: {
-          getCurrentPrice: stores.coinPriceStore.getCurrentPrice,
         },
         addresses: {
           addressSubgroupMap: stores.addresses.addressSubgroupMap,
@@ -308,11 +139,6 @@ export default class UnmangleTxDialogContainer extends Component<Props> {
           },
           reset: {
             trigger: actions.txBuilderActions.reset.trigger
-          },
-        },
-        wallets: {
-          sendMoney: {
-            trigger: actions.wallets.sendMoney.trigger
           },
         },
       },
