+++ stories/helpers/StoryWrapper.js
@@ -19,18 +19,6 @@ import es from 'react-intl/locale-data/es';
 import it from 'react-intl/locale-data/it';
 import tr from 'react-intl/locale-data/tr';
 import '../../app/themes/index.global.scss';
-import type { UnitOfAccountSettingType } from '../../app/types/unitOfAccountType';
-import type { V3UnsignedTxAddressedUtxoResponse } from '../../app/api/ada/transactions/types';
-import type { ISignRequest } from '../../app/api/common/lib/transactions/ISignRequest';
-import {
-  genToAbsoluteSlotNumber,
-  genToRelativeSlotNumber,
-  genTimeToSlot,
-  genCurrentEpochLength,
-  genCurrentSlotLength,
-  genTimeSinceGenesis,
-  genToRealTime,
-} from '../../app/api/jormungandr/lib/storage/bridge/timeUtils';
 import { yoroiPolymorphTheme } from '../../app/themes/PolymorphThemes';
 import { themeOverrides } from '../../app/themes/overrides';
 import { translations, LANGUAGES } from '../../app/i18n/translations';
@@ -39,49 +27,14 @@ import { THEMES, changeToplevelTheme } from '../../app/themes';
 import type { Theme } from '../../app/themes';
 import environment from '../../app/environment';
 import { getVarsForTheme } from '../../app/stores/toplevel/ProfileStore';
-import type { HwWalletMetaRow, } from '../../app/api/ada/lib/storage/database/walletTypes/core/tables';
-import { assuranceModes } from '../../app/config/transactionAssuranceConfig';
 
 import { withKnobs, select, boolean } from '@storybook/addon-knobs';
 import { addDecorator } from '@storybook/react';
 
-import { PublicDeriver } from '../../app/api/ada/lib/storage/models/PublicDeriver';
-import { Cip1852Wallet } from '../../app/api/ada/lib/storage/models/Cip1852Wallet/wrapper';
-import { Bip44Wallet } from '../../app/api/ada/lib/storage/models/Bip44Wallet/wrapper';
-import { WalletTypeOption } from '../../app/api/ada/lib/storage/models/ConceptualWallet/interfaces';
-import {
-  HasPrivateDeriver,
-  HasSign,
-  HasLevels,
-  GetAllUtxos,
-  GetSigningKey,
-  GetPublicKey,
-  DisplayCutoff,
-  Cip1852JormungandrPickInternal,
-  Bip44PickInternal,
-  GetAllAccounting,
-  GetStakingKey,
-  HasUtxoChains,
-} from '../../app/api/ada/lib/storage/models/PublicDeriver/traits';
-import {
-  Bip44DerivationLevels,
-} from '../../app/api/ada/lib/storage/database/walletTypes/bip44/api/utils';
-import type { ConceptualWalletSettingsCache } from '../../app/stores/toplevel/WalletSettingsStore';
-import WalletSettingsStore from '../../app/stores/toplevel/WalletSettingsStore';
-import TransactionsStore from '../../app/stores/toplevel/TransactionsStore';
-import DelegationStore from '../../app/stores/toplevel/DelegationStore';
-import WalletStore from '../../app/stores/toplevel/WalletStore';
-import AdaTimeStore from '../../app/stores/ada/AdaTimeStore';
-import CachedRequest from '../../app/stores/lib/LocalizedCachedRequest';
 import LocalizableError from '../../app/i18n/LocalizableError';
 import globalMessages from '../../app/i18n/global-messages';
 import { ledgerErrors } from '../../app/domain/LedgerLocalizedError';
-import BigNumber from 'bignumber.js';
-import { utxoToTxInput } from '../../app/api/jormungandr/lib/transactions/inputSelection';
-import { RustModule } from '../../app/api/ada/lib/cardanoCrypto/rustLoader';
-import { networks, getJormungandrBaseConfig, isCardanoHaskell, getCardanoHaskellBaseConfig, isJormungandr } from '../../app/api/ada/lib/storage/database/prepackaged/networks';
-import { HaskellShelleyTxSignRequest } from '../../app/api/ada/transactions/shelley/HaskellShelleyTxSignRequest';
-import { JormungandrTxSignRequest } from '../../app/api/jormungandr/lib/transactions/JormungandrTxSignRequest';
+import type { UnitOfAccountSettingType } from '../../app/types/unitOfAccountType';
 
 /**
  * This whole file is meant to mirror code in App.js
@@ -305,518 +258,6 @@ export const mockTrezorMeta = {
   DeviceId: 'C875BA9D0C571FF4B8718FAA',
   Model: 'T',
   Vendor: 'trezor.io',
-};
-
-let conceptualWalletCounter = 0;
-let publicDeriverCounter = 0;
-
-function genDummyWallet(): PublicDeriver<> {
-  const conceptualWalletId = conceptualWalletCounter++;
-  const parent = new Cip1852Wallet(
-    (null: any),
-    {
-      db: (null: any),
-      conceptualWalletId,
-      walletType: WalletTypeOption.WEB_WALLET,
-      hardwareInfo: null,
-      networkInfo: networks.ByronMainnet,
-    },
-    {
-      ConceptualWalletId: conceptualWalletId,
-      SignerLevel: null,
-      PublicDeriverLevel: 0,
-      PrivateDeriverLevel: null,
-      PrivateDeriverKeyDerivationId: null,
-      RootKeyDerivationId: 0,
-    },
-    null,
-    null,
-  );
-  const clazz = GetPublicKey(HasLevels(HasSign(PublicDeriver)));
-  const self = new clazz({
-    publicDeriverId: publicDeriverCounter++,
-    parent,
-    pathToPublic: [],
-    derivationId: 0,
-  });
-  return self;
-}
-
-function genMockCache(dummyWallet: PublicDeriver<>) {
-  const pendingRequest = new CachedRequest(_publicDeriver => Promise.resolve([]));
-  const recentRequest = new CachedRequest(_request => Promise.resolve({
-    transactions: [],
-    total: 0,
-  }));
-  const allRequest = new CachedRequest(_request => Promise.resolve({
-    transactions: [],
-    total: 0,
-  }));
-  const getBalanceRequest = new CachedRequest(_request => Promise.resolve(
-    new BigNumber(0),
-  ));
-  return {
-    conceptualWalletCache: {
-      conceptualWallet: dummyWallet.getParent(),
-      conceptualWalletName: 'Test wallet',
-    },
-    getPublicKeyCache: (wallet) => ({
-      publicDeriver: wallet,
-      plate: {
-        ImagePart: '8e4e2f11b6ac2a269913286e26339779ab8767579d18d173cdd324929d94e2c43e3ec212cc8a36ed9860579dfe1e3ef4d6de778c5dbdd981623b48727cd96247',
-        TextPart: 'DNKO-8098',
-      },
-    }),
-    getDelegation: (_wallet) => (undefined),
-    getTransactions: (wallet) => ({
-      publicDeriver: wallet,
-      lastSyncInfo: {
-        BlockHash: null,
-        LastSyncInfoId: 1,
-        SlotNum: null,
-        Height: 0,
-        Time: null,
-      },
-      requests: {
-        pendingRequest,
-        recentRequest,
-        allRequest,
-        getBalanceRequest,
-      },
-    }),
-    getPublicDeriverSettingsCache: (publicDeriver) => ({
-      publicDeriver,
-      assuranceMode: assuranceModes.NORMAL,
-      publicDeriverName: '',
-    }),
-    getSigningKeyCache: (publicDeriver) => ({
-      publicDeriver,
-      signingKeyUpdateDate: null,
-    }),
-    getTimeCalcRequests: (publicDeriver) => ({
-      publicDeriver,
-      requests: {
-        toAbsoluteSlot: new CachedRequest(() => {
-          return genToAbsoluteSlotNumber(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        toRelativeSlotNumber: new CachedRequest(() => {
-          return genToRelativeSlotNumber(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        timeToSlot: new CachedRequest(() => {
-          return genTimeToSlot(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        currentEpochLength: new CachedRequest(() => {
-          return genCurrentEpochLength(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        currentSlotLength: new CachedRequest(() => {
-          return genCurrentSlotLength(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        timeSinceGenesis: new CachedRequest(() => {
-          return genTimeSinceGenesis(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-        toRealTime: new CachedRequest(() => {
-          return genToRealTime(
-            getJormungandrBaseConfig(publicDeriver.getParent().getNetworkInfo())
-          );
-        }),
-      },
-    }),
-    getCurrentTimeRequests: (publicDeriver) => ({
-      publicDeriver,
-      currentEpoch: 100,
-      currentSlot: 5000,
-      msIntoSlot: 10,
-    }),
-  };
-}
-
-export function genDummyWithCache(): CacheValue {
-  const dummyWallet = genDummyWallet();
-  return {
-    publicDeriver: dummyWallet,
-    ...genMockCache(dummyWallet),
-  };
-}
-
-function genSigningWallet(
-  genHardwareInfo?: number => HwWalletMetaRow,
-): PublicDeriver<> {
-  const conceptualWalletId = conceptualWalletCounter++;
-  const hardwareInfo = genHardwareInfo == null
-    ? null
-    : genHardwareInfo(conceptualWalletId);
-  const parent = new Cip1852Wallet(
-    (null: any),
-    {
-      db: (null: any),
-      conceptualWalletId,
-      walletType: (() => {
-        if (hardwareInfo != null) {
-          return WalletTypeOption.HARDWARE_WALLET;
-        }
-        return WalletTypeOption.WEB_WALLET;
-      })(),
-      hardwareInfo,
-      networkInfo: networks.JormungandrMainnet,
-    },
-    {
-      ConceptualWalletId: conceptualWalletId,
-      SignerLevel: null,
-      PublicDeriverLevel: Bip44DerivationLevels.ACCOUNT.level,
-      PrivateDeriverLevel: null,
-      PrivateDeriverKeyDerivationId: null,
-      RootKeyDerivationId: 0,
-    },
-    null,
-    null,
-  );
-  const clazz = HasUtxoChains(Cip1852JormungandrPickInternal(GetStakingKey(GetAllAccounting(
-    DisplayCutoff(GetSigningKey(GetPublicKey(
-      GetAllUtxos(HasLevels(HasSign(HasPrivateDeriver((PublicDeriver: any)))))
-    )))
-  ))));
-  const self = new clazz({
-    publicDeriverId: publicDeriverCounter++,
-    parent,
-    pathToPublic: [],
-    derivationId: 0,
-  });
-  return self;
-}
-function genByronSigningWallet(
-  genHardwareInfo?: number => HwWalletMetaRow,
-): PublicDeriver<> {
-  const conceptualWalletId = conceptualWalletCounter++;
-  const hardwareInfo = genHardwareInfo == null
-    ? null
-    : genHardwareInfo(conceptualWalletId);
-  const parent = new Bip44Wallet(
-    (null: any),
-    {
-      db: (null: any),
-      conceptualWalletId,
-      walletType: (() => {
-        if (hardwareInfo != null) {
-          return WalletTypeOption.HARDWARE_WALLET;
-        }
-        return WalletTypeOption.WEB_WALLET;
-      })(),
-      hardwareInfo,
-      networkInfo: networks.ByronMainnet,
-    },
-    {
-      ConceptualWalletId: conceptualWalletId,
-      SignerLevel: null,
-      PublicDeriverLevel: Bip44DerivationLevels.ACCOUNT.level,
-      PrivateDeriverLevel: null,
-      PrivateDeriverKeyDerivationId: null,
-      RootKeyDerivationId: 0,
-    },
-    null,
-    null,
-  );
-  const clazz = HasUtxoChains(Bip44PickInternal(
-    DisplayCutoff(GetSigningKey(GetPublicKey(
-      GetAllUtxos(HasLevels(HasSign(HasPrivateDeriver((PublicDeriver: any)))))
-    )))
-  ));
-  const self = new clazz({
-    publicDeriverId: publicDeriverCounter++,
-    parent,
-    pathToPublic: [],
-    derivationId: 0,
-  });
-  return self;
-}
-
-export function genSigningWalletWithCache(
-  genHardwareInfo?: number => HwWalletMetaRow,
-): CacheValue {
-  const dummyWallet = genSigningWallet(genHardwareInfo);
-  return {
-    publicDeriver: dummyWallet,
-    ...genMockCache(dummyWallet),
-  };
-}
-export function getByronSigningWalletWithCache(
-  genHardwareInfo?: number => HwWalletMetaRow,
-): CacheValue {
-  const dummyWallet = genByronSigningWallet(genHardwareInfo);
-  return {
-    publicDeriver: dummyWallet,
-    ...genMockCache(dummyWallet),
-  };
-}
-
-export type CacheValue = {|
-  publicDeriver: PublicDeriver<>,
-  conceptualWalletCache: ConceptualWalletSettingsCache,
-  getPublicKeyCache:
-    typeof WalletStore.prototype.getPublicKeyCache,
-  getTransactions:
-    typeof TransactionsStore.prototype.getTxRequests,
-  getDelegation:
-    typeof DelegationStore.prototype.getDelegationRequests,
-  getSigningKeyCache:
-    typeof WalletStore.prototype.getSigningKeyCache,
-  getPublicDeriverSettingsCache:
-    typeof WalletSettingsStore.prototype.getPublicDeriverSettingsCache,
-  getTimeCalcRequests:
-    typeof AdaTimeStore.prototype.getTimeCalcRequests,
-  getCurrentTimeRequests:
-    typeof AdaTimeStore.prototype.getCurrentTimeRequests,
-|};
-
-export function walletLookup(wallets: Array<CacheValue>): {|
-  publicDerivers: Array<PublicDeriver<>>,
-  getConceptualWalletSettingsCache:
-    typeof WalletSettingsStore.prototype.getConceptualWalletSettingsCache,
-  getPublicKeyCache:
-    typeof WalletStore.prototype.getPublicKeyCache,
-  getTransactions:
-    typeof TransactionsStore.prototype.getTxRequests,
-  getDelegation:
-    typeof DelegationStore.prototype.getDelegationRequests,
-  getSigningKeyCache:
-    typeof WalletStore.prototype.getSigningKeyCache,
-  getPublicDeriverSettingsCache:
-    typeof WalletSettingsStore.prototype.getPublicDeriverSettingsCache,
-  getTimeCalcRequests:
-    typeof AdaTimeStore.prototype.getTimeCalcRequests,
-  getCurrentTimeRequests:
-    typeof AdaTimeStore.prototype.getCurrentTimeRequests,
-|} {
-  if (wallets.length === 0) {
-    return ({
-      publicDerivers: [],
-      getConceptualWalletSettingsCache: (_conceptualWallet) => (null: any),
-      getTransactions: (_publicDeriver) => (null: any),
-      getDelegation: (_publicDeriver) => (null: any),
-      getPublicKeyCache: (_publicDeriver) => (null: any),
-      getSigningKeyCache: (_publicDeriver) => (null: any),
-      getPublicDeriverSettingsCache: (_publicDeriver) => (null: any),
-      getTimeCalcRequests: (_publicDeriver) => (null: any),
-      getCurrentTimeRequests: (_publicDeriver) => (null: any),
-    });
-  }
-
-  const asOption: { [key: string]: PublicDeriver<>, ... } = {};
-  for (const wallet of wallets) {
-    asOption[wallet.conceptualWalletCache.conceptualWalletName] = wallet.publicDeriver;
-  }
-
-  return ({
-    publicDerivers: wallets.map(wallet => wallet.publicDeriver),
-    getConceptualWalletSettingsCache: (conceptualWallet) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver.getParent() === conceptualWallet) {
-          return wallet.conceptualWalletCache;
-        }
-      }
-      throw new Error(`Missing cache entry for getConceptualWalletSettingsCache`);
-    },
-    getTransactions: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getTransactions(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for transactions`);
-    },
-    getDelegation: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getDelegation(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for delegation`);
-    },
-    getPublicKeyCache: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getPublicKeyCache(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for publicKeyCache`);
-    },
-    getSigningKeyCache: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getSigningKeyCache(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for getSigningKeyCache`);
-    },
-    getPublicDeriverSettingsCache: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getPublicDeriverSettingsCache(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for getPublicDeriverSettingsCache`);
-    },
-    getTimeCalcRequests: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getTimeCalcRequests(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for getTimeCalcRequests`);
-    },
-    getCurrentTimeRequests: (publicDeriver) => {
-      for (const wallet of wallets) {
-        if (wallet.publicDeriver === publicDeriver) {
-          return wallet.getCurrentTimeRequests(publicDeriver);
-        }
-      }
-      throw new Error(`Missing cache entry for getCurrentTimeRequests`);
-    },
-  });
-}
-
-export const genTentativeTx = (
-  publicDeriver: PublicDeriver<>,
-): {|
-  tentativeTx: null | ISignRequest<any>,
-  inputAmount: string,
-  fee: BigNumber,
-|} => {
-  const inputAmount = '1000001';
-  const ouputAmount = '400';
-  const fee = new BigNumber(inputAmount).minus(new BigNumber(ouputAmount));
-
-  const networkInfo = publicDeriver.getParent().getNetworkInfo();
-  if (isJormungandr(networkInfo)) {
-    const remoteUnspentUtxo = {
-      amount: inputAmount,
-      receiver: 'Ae2tdPwUPEZKX8N2TjzBXLy5qrecnQUniTd2yxE8mWyrh2djNpUkbAtXtP4',
-      tx_hash: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe',
-      tx_index: 0,
-      utxo_id: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe0',
-    };
-    const input = utxoToTxInput(remoteUnspentUtxo);
-    const builder = RustModule.WalletV3.InputOutputBuilder.empty();
-    builder.add_input(input);
-    builder.add_output(
-      RustModule.WalletV3.Address.from_string('addr1s33chdhaexquujgnwm458swlt4tl2t5qyyk7d04gtdy5utp4y6c9sf4nrtlu7glkdwww3leg94jr6rkt9prwfgfp7symp5nj08zuln4lmxjv7k'),
-      RustModule.WalletV3.Value.from_str(ouputAmount)
-    );
-    const unsignedTx = builder.build();
-    return {
-      tentativeTx: new JormungandrTxSignRequest({
-        senderUtxos: [{
-          ...remoteUnspentUtxo,
-          addressing: {
-            path: [],
-            startLevel: 0,
-          },
-        }],
-        unsignedTx,
-        changeAddr: [],
-        certificate: undefined,
-      }),
-      inputAmount,
-      fee,
-    };
-  }
-  if (isCardanoHaskell(networkInfo)) {
-    const config = getCardanoHaskellBaseConfig(networkInfo)
-      .reduce((acc, next) => Object.assign(acc, next), {});
-    const remoteUnspentUtxo = {
-      amount: inputAmount,
-      receiver: 'Ae2tdPwUPEZKX8N2TjzBXLy5qrecnQUniTd2yxE8mWyrh2djNpUkbAtXtP4',
-      tx_hash: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe',
-      tx_index: 0,
-      utxo_id: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe0',
-    };
-    const txBuilder = RustModule.WalletV4.TransactionBuilder.new(
-      RustModule.WalletV4.LinearFee.new(
-        RustModule.WalletV4.BigNum.from_str(config.LinearFee.coefficient),
-        RustModule.WalletV4.BigNum.from_str(config.LinearFee.constant),
-      ),
-      RustModule.WalletV4.BigNum.from_str(config.MinimumUtxoVal),
-      RustModule.WalletV4.BigNum.from_str(config.PoolDeposit),
-      RustModule.WalletV4.BigNum.from_str(config.KeyDeposit),
-    );
-    txBuilder.add_bootstrap_input(
-      RustModule.WalletV4.ByronAddress.from_base58(remoteUnspentUtxo.receiver),
-      RustModule.WalletV4.TransactionInput.new(
-        RustModule.WalletV4.TransactionHash.from_bytes(
-          Buffer.from(remoteUnspentUtxo.tx_hash, 'hex')
-        ),
-        remoteUnspentUtxo.tx_index
-      ),
-      RustModule.WalletV4.BigNum.from_str(remoteUnspentUtxo.amount)
-    );
-    txBuilder.add_output(RustModule.WalletV4.TransactionOutput.new(
-      RustModule.WalletV4.ByronAddress.from_base58('Ae2tdPwUPEZ4xAL3nxLq4Py7BfS1D2tJ3u2rxZGnrAXC8TNkWhTaz41J3FN').to_address(),
-      RustModule.WalletV4.BigNum.from_str(ouputAmount)
-    ));
-    txBuilder.set_fee(RustModule.WalletV4.BigNum.from_str('5'));
-    txBuilder.set_ttl(5);
-    return {
-      tentativeTx: new HaskellShelleyTxSignRequest({
-        senderUtxos: [{
-          ...remoteUnspentUtxo,
-          addressing: {
-            path: [],
-            startLevel: 0,
-          },
-        }],
-        unsignedTx: txBuilder,
-        changeAddr: [],
-        certificate: undefined,
-      }),
-      inputAmount,
-      fee,
-    };
-  }
-  throw new Error(`Unknown network ${JSON.stringify(networkInfo)}`);
-};
-
-export const genUndelegateTx = (
-  publicDeriver: PublicDeriver<>,
-): JormungandrTxSignRequest => {
-  const inputAmount = '1000001';
-
-  if (!isJormungandr(publicDeriver.getParent().getNetworkInfo())) {
-    throw new Error('Delegation only supported for Jormungandr');
-  }
-  const remoteUnspentUtxo = {
-    amount: inputAmount,
-    receiver: 'Ae2tdPwUPEZKX8N2TjzBXLy5qrecnQUniTd2yxE8mWyrh2djNpUkbAtXtP4',
-    tx_hash: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe',
-    tx_index: 0,
-    utxo_id: '6930f123df83e4178b0324ae617b2028c0b38c6ff4660583a2abf1f7b08195fe0',
-  };
-  const input = utxoToTxInput(remoteUnspentUtxo);
-  const builder = RustModule.WalletV3.InputOutputBuilder.empty();
-  builder.add_input(input);
-  const IOs = builder.build();
-  return new JormungandrTxSignRequest({
-    senderUtxos: [{
-      ...remoteUnspentUtxo,
-      addressing: {
-        path: [],
-        startLevel: 0,
-      },
-    }],
-    unsignedTx: IOs,
-    changeAddr: [],
-    certificate: undefined, // TODO
-  });
 };
 
 export const genUnitOfAccount: void => UnitOfAccountSettingType = () => {
